requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.1.2=pypi_0
annotated-types=0.6.0=pypi_0
anyio=3.7.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
azure-cognitiveservices-vision-computervision=0.9.0=pypi_0
azure-common=1.1.28=pypi_0
azure-core=1.29.4=pypi_0
backports-zoneinfo=0.2.1=pypi_0
blinker=1.6.3=pypi_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.0=pypi_0
click=8.1.7=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
dataclasses-json=0.6.1=pypi_0
editdistance=0.6.2=pypi_0
efficientnet=1.0.0=pypi_0
essential-generators=1.0=pypi_0
exceptiongroup=1.1.3=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.43.1=pypi_0
frozenlist=1.4.0=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.37=pypi_0
google-auth=2.23.2=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
greenlet=3.0.0=pypi_0
grpcio=1.59.0=pypi_0
h5py=3.9.0=pypi_0
idna=3.4=pypi_0
imageio=2.31.5=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.1.0=pypi_0
isodate=0.6.1=pypi_0
jinja2=3.1.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.1=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
keras=2.13.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-ocr=0.0.0=pypi_0
kiwisolver=1.4.5=pypi_0
langchain=0.0.310=pypi_0
langsmith=0.0.43=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.6=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.5=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.20.1=pypi_0
matplotlib=3.7.3=pypi_0
mdurl=0.1.2=pypi_0
msrest=0.7.1=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
openai=0.28.1=pypi_0
opencv-python=4.8.1.78=pypi_0
openssl=3.0.11=h7f8727e_2
opt-einsum=3.3.0=pypi_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
pillow=10.0.1=pypi_0
pip=23.2.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
protobuf=4.24.4=pypi_0
pyarrow=13.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyclipper=1.3.0.post5=pypi_0
pydantic=2.4.2=pypi_0
pydantic-core=2.10.1=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.16.1=pypi_0
pyparsing=3.1.1=pypi_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.6.0=pypi_0
rpds-py=0.10.4=pypi_0
rsa=4.9=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.0.0=py38h06a4308_0
shapely=2.0.1=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
sqlalchemy=2.0.21=pypi_0
sqlite=3.41.2=h5eee18b_0
streamlit=1.27.2=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow=2.13.1=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
termcolor=2.3.0=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.8.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=5.1=pypi_0
urllib3=2.0.6=pypi_0
validators=0.22.0=pypi_0
watchdog=3.0.0=pypi_0
werkzeug=3.0.0=pypi_0
wheel=0.41.2=py38h06a4308_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
